+++ examples/f413disco-lcd-ferris.rs
@@ -16,7 +16,7 @@ use rtt_target::{self, rtt_init_print};
 use stm32f4xx_hal as hal;
 
 use crate::hal::{
-    fsmc_lcd::{FsmcLcd, LcdPins, Timing},
+    fsmc_lcd::{DataPins16, FsmcLcd, LcdPins, Timing},
     gpio::Speed,
     pac::{CorePeripherals, Peripherals},
     prelude::*,
@@ -710,30 +710,17 @@ fn main() -> ! {
 
         // Define the pins we need for our 16bit parallel bus
         use stm32f4xx_hal::gpio::alt::fsmc as alt;
-        let lcd_pins = LcdPins {
-            data: (
-                gpiod.pd14.into(),
-                gpiod.pd15.into(),
-                gpiod.pd0.into(),
-                gpiod.pd1.into(),
-                gpioe.pe7.into(),
-                gpioe.pe8.into(),
-                gpioe.pe9.into(),
-                gpioe.pe10.into(),
-                gpioe.pe11.into(),
-                gpioe.pe12.into(),
-                gpioe.pe13.into(),
-                gpioe.pe14.into(),
-                gpioe.pe15.into(),
-                gpiod.pd8.into(),
-                gpiod.pd9.into(),
-                gpiod.pd10.into(),
+        let lcd_pins = LcdPins::new(
+            DataPins16::new(
+                gpiod.pd14, gpiod.pd15, gpiod.pd0, gpiod.pd1, gpioe.pe7, gpioe.pe8, gpioe.pe9,
+                gpioe.pe10, gpioe.pe11, gpioe.pe12, gpioe.pe13, gpioe.pe14, gpioe.pe15, gpiod.pd8,
+                gpiod.pd9, gpiod.pd10,
             ),
-            address: alt::Address::from(gpiof.pf0),
-            read_enable: gpiod.pd4.into(),
-            write_enable: gpiod.pd5.into(),
-            chip_select: alt::ChipSelect3::from(gpiog.pg10),
-        };
+            alt::Address::from(gpiof.pf0),
+            gpiod.pd4,
+            gpiod.pd5,
+            alt::ChipSelect3::from(gpiog.pg10),
+        );
 
         // Setup the RESET pin
         let rst = gpiob.pb13.into_push_pull_output().speed(Speed::VeryHigh);