+++ src/gpio/alt.rs
@@ -66,7 +66,7 @@ use extipin;
 
 macro_rules! pin {
     ( $($(#[$docs:meta])* <$name:ident, $Otype:ident> for $(no: $NoPin:ident,)? [$(
-        $(#[$attr:meta])* $PX:ident<$A:literal>,
+        $(#[$attr:meta])* $PX:ident<$A:literal $(, Speed::$Speed:ident)?>,
     )*],)*) => {
         $(
             #[derive(Debug)]
@@ -143,14 +143,14 @@ macro_rules! pin {
                     MODE: $crate::gpio::marker::NotAlt + $crate::gpio::PinMode
                 {
                     fn from(p: gpio::$PX<MODE>) -> Self {
-                        Self::$PX(p.into_mode())
+                        Self::$PX(p.into_mode() $(.speed($crate::gpio::Speed::$Speed))?)
                     }
                 }
 
                 $(#[$attr])*
                 impl From<gpio::$PX<$crate::gpio::Alternate<$A, $Otype>>> for $name {
                     fn from(p: gpio::$PX<$crate::gpio::Alternate<$A, $Otype>>) -> Self {
-                        Self::$PX(p)
+                        Self::$PX(p $(.speed($crate::gpio::Speed::$Speed))?)
                     }
                 }